Punkty kontrolne i drony

+++ src/dataModels/Dron.java
@@ -6,25 +6,25 @@ import javax.xml.ws.soap.Addressing;
 import java.lang.annotation.Target;
 
 public class Dron extends DataModel{
-  private Long id;
+  private Integer id;
   private String nazwa;
   private String opis;
-  private Double masa;
-  private Long ilosc_wirnikow;
-  private Double max_predkosc;
-  private Double max_czas_lotu;
-  private Double poziom_baterii;
-  private Double wspx;
-  private Double wspy;
-  private Double wspz;
-  private Long stan;
-  private Long punkt_kontrolny_id;
+  private Float masa;
+  private Integer ilosc_wirnikow;
+  private Float max_predkosc;
+  private Float max_czas_lotu;
+  private Float poziom_baterii;
+  private Float wspx;
+  private Float wspy;
+  private Float wspz;
+  private Integer stan;
+  private Integer punkt_kontrolny_id;
 
-  public Long getId() {
+  public Integer getId() {
     return id;
   }
 
-  public void setId(Long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
 
@@ -32,9 +32,7 @@ public class Dron extends DataModel{
     return nazwa;
   }
 
-  public void setNazwa(String nazwa) {
-    this.nazwa = nazwa;
-  }
+  public void setNazwa(String nazwa) { this.nazwa = nazwa; }
 
   public String getOpis() {
     return opis;
@@ -44,83 +42,83 @@ public class Dron extends DataModel{
     this.opis = opis;
   }
 
-  public Double getMasa() {
+  public Float getMasa() {
     return masa;
   }
 
-  public void setMasa(Double masa) {
+  public void setMasa(Float masa) {
     this.masa = masa;
   }
 
-  public Long getIlosc_wirnikow() {
+  public Integer getIlosc_wirnikow() {
     return ilosc_wirnikow;
   }
 
-  public void setIlosc_wirnikow(Long ilosc_wirnikow) {
+  public void setIlosc_wirnikow(Integer ilosc_wirnikow) {
     this.ilosc_wirnikow = ilosc_wirnikow;
   }
 
-  public Double getMax_predkosc() {
+  public Float getMax_predkosc() {
     return max_predkosc;
   }
 
-  public void setMax_predkosc(Double max_predkosc) {
+  public void setMax_predkosc(Float max_predkosc) {
     this.max_predkosc = max_predkosc;
   }
 
-  public Double getMax_czas_lotu() {
+  public Float getMax_czas_lotu() {
     return max_czas_lotu;
   }
 
-  public void setMax_czas_lotu(Double max_czas_lotu) {
+  public void setMax_czas_lotu(Float max_czas_lotu) {
     this.max_czas_lotu = max_czas_lotu;
   }
 
-  public Double getPoziom_baterii() {
+  public Float getPoziom_baterii() {
     return poziom_baterii;
   }
 
-  public void setPoziom_baterii(Double poziom_baterii) {
+  public void setPoziom_baterii(Float poziom_baterii) {
     this.poziom_baterii = poziom_baterii;
   }
 
-  public Double getWspx() {
+  public Float getWspx() {
     return wspx;
   }
 
-  public void setWspx(Double wspx) {
+  public void setWspx(Float wspx) {
     this.wspx = wspx;
   }
 
-  public Double getWspy() {
+  public Float getWspy() {
     return wspy;
   }
 
-  public void setWspy(Double wspy) {
+  public void setWspy(Float wspy) {
     this.wspy = wspy;
   }
 
-  public Double getWspz() {
+  public Float getWspz() {
     return wspz;
   }
 
-  public void setWspz(Double wspz) {
+  public void setWspz(Float wspz) {
     this.wspz = wspz;
   }
 
-  public Long getStan() {
+  public Integer getStan() {
     return stan;
   }
 
-  public void setStan(Long stan) {
+  public void setStan(Integer stan) {
     this.stan = stan;
   }
 
-  public Long getPunkt_kontrolny_id() {
+  public Integer getPunkt_kontrolny_id() {
     return punkt_kontrolny_id;
   }
 
-  public void setPunkt_kontrolny_id(Long punkt_kontrolny_id) {
+  public void setPunkt_kontrolny_id(Integer punkt_kontrolny_id) {
     this.punkt_kontrolny_id = punkt_kontrolny_id;
   }
 }
